Fix: Rename package to hello_world_mvp and fetch Google authCode
