+++ app/src/androidTest/java/com/example/hello_world_mvp/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.helloworld
+package com.example.hello_world_mvp
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
