+++ app/src/main/java/com/example/hello_world_mvp/HelloWorldApplication.kt
@@ -1,4 +1,4 @@
-package com.example.helloworld
+package com.example.hello_world_mvp
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
