+++ app/src/test/java/com/example/hello_world_mvp/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.helloworld
+package com.example.hello_world_mvp
 
 import org.junit.Test
